refactor: change folders name and update imports
