+++ src/components/clothingGrid/DescriptionWrapper.tsx
